SCRUM-6: Isteyak fixed full service request
