--- python/first_10_requests.json
@@ -1,10 +1,0 @@
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"6##","Street":"JANETTE","Ward":"WARD 3","Method Received":"Citizen Mobile Apps","Created Date":"JUN 24, 2024 03:26:22 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"12##","Street":"LABOUR","Ward":"WARD 9","Method Received":"Phone","Created Date":"JUN 24, 2024 03:28:28 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"17##","Street":"VICKY","Ward":"WARD 9","Method Received":"Text","Created Date":"JUN 24, 2024 03:58:37 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"98##","Street":"LYNNGROVE","Ward":"WARD 7","Method Received":"E-Mail","Created Date":"JUN 25, 2024 09:04:31 AM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"1##","Street":"OAK","Ward":"WARD 3","Method Received":"Phone","Created Date":"JUN 25, 2024 01:16:21 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"34##","Street":"CARIBOU","Ward":"WARD 9","Method Received":"Citizen Mobile Apps","Created Date":"JUN 27, 2024 10:27:01 AM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"22##","Street":"ASKIN","Ward":"WARD 10","Method Received":"Phone","Created Date":"JUL 02, 2024 01:40:11 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"9##","Street":"ELSMERE","Ward":"WARD 4","Method Received":"Phone","Created Date":"AUG 27, 2024 01:13:40 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"31##","Street":"TURNER","Ward":"WARD 9","Method Received":"Phone","Created Date":"AUG 27, 2024 01:47:54 PM"}
-{"Service Request Description":"3 Day Parking Infraction","Department":"Parking Enforcement","Block\/Address":"10##","Street":"ASKIN","Ward":"WARD 2","Method Received":"Phone","Created Date":"AUG 27, 2024 12:03:12 PM"}
